LayoutTests/http/tests/storageAccess/deny-storage-access-under-opener-expected.txt

Tests that a cross-origin window from a prevalent domain without user interaction is denied storage access under its opener.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Cookie created.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Should not receive first-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named ''.
Client-side document.cookie:
